Refactor router
To make lib.rs, and the router in particular more readable, move
structs and methods into separate data module. Expand data module to
capture more of the data handled by the router and the functions called
within.
This makes the router much easier to follow, and reduces control flow
complexity.